pyproject.toml
# quickstart:
#   pdm install
#   pdm run pytest

# dependencies
#   - use `pdm add <package>` and `pdm remove <package>` to manage dependencies
#   - use `pdm add -d -G dev <packag>` to add a dev dependency
#   - use `pdm update --update-all` to update dependencies
#   - use `pdm install` to install dependencies

# usage:
#   - use `pdm run <pythonfile>` to execute code
#   - use `pdm run pytest` to perform unit tests

# line lengths:
#   - Excessive line lengths aren't good. But rather then automatically
#     breaking them and turning into vertical blocks we want to see the
#     nastiness sticking out prominently.

[project]
name = "feeph.github_utils"
version = "0.0.0"
description = "various GitHub related utilities"
authors = [
    {name = "Feeph Aifeimei", email = "55798703+feeph@users.noreply.github.com"},
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "attrs~=24.2",
    "pynacl~=1.5",
    "requests~=2.32",
    "pyyaml~=6.0",
]
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
repository = "https://github.com/feeph/github-utils-python/"
keywords = [
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Typing :: Typed",
]

[project.scripts]
update_github-secrets = "feeph.github_utils.update_secrets:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

# -------------------------------------------------------------------------
# tooling
# -------------------------------------------------------------------------

# https://github.com/PyCQA/isort
# https://pycqa.github.io/isort/
[tool.isort]
# use the black profile but override "multi_line_output"
profile           = "black"
line_length       = 250
multi_line_output = 7
src_paths         = ["feeph/", "tests/"]

# https://pypi.org/project/autopep8/
[tool.autopep8]
max_line_length = 250
ignore          = ["E221", "E241"]
in-place = true

[tool.pdm]
distribution = true

# automatically update requirements.txt and requirements-dev.txt
# on `pdm lock`

[[tool.pdm.autoexport]]
filename = "requirements.txt"
groups = ["default"]

[[tool.pdm.autoexport]]
filename = "requirements-dev.txt"
groups = ["dev", "tools"]

[tool.pdm.build]
includes = ["feeph/"]
source-includes = ["tests/"]

[tool.pdm.dev-dependencies]
dev = [
    "types-PyYAML~=6.0",
    "types-requests~=2.32",
]
tools = [
    "autopep8      ~= 2.2",
    "copier        ~= 9.3",
    "coverage-lcov ~= 0.3",
    "flake8        ~= 7.0",
    "mypy          ~= 1.10",
    "pylint        ~= 3.2",
    "pytest-cov    ~= 5.0",
    "pytest-sugar  ~= 1.0",
]

# during development: fetch all packages in our namespace from TestPyPI
#
# this config is used exclusively by pdm and has no impact on regular
# consumers using `pip install <package>` to install the package
[[tool.pdm.source]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
include_packages = ["feeph-*"]

[tool.pylint."MESSAGES CONTROL"]
fail-under = 8
max-line-length = 250
disable = "no-else-return,use-list-literal"
